Heroku migration (#18)

+++ src/main/java/pro/belbix/ethparser/dto/HarvestDTO.java
@@ -23,7 +23,7 @@ public class HarvestDTO implements DtoI {
     @Id
     private String id;
     private String hash;
-    private BigInteger block;
+    private Long block;
     private boolean confirmed = false;
     private Long blockDate;
     private String methodName;
